requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paho-mqtt==1.6.1
pycparser==2.21
Werkzeug==2.2.3
